fix: 찜 해제 시 순서 변경되는 문제 해결

+++ android/app/src/main/java/com/on/turip/data/place/dto/FavoritePlaceResponse.kt
@@ -10,5 +10,5 @@ data class FavoritePlaceResponse(
     @SerialName("place")
     val placeResponse: PlaceResponse,
     @SerialName("favoriteOrder")
-    val order: Int,
+    val order: Long,
 )
